updated_with_new_seller_module_01_06_2023_10_50

+++ app/src/main/java/com/project/motoassistant/models/WorkshopDetail.java
@@ -12,7 +12,7 @@ public class WorkshopDetail {
     public String password;
     public String file;
     public String distance;
-    public int rat_percentage;
+    public double rat_percentage;
     public double workshop_rating;
     public int no_of_rating;
     public String created_at;
